requirements.txt
httpx==0.21.1
mysql-connector-python==8.0.27
beautifulsoup4==4.10.0
